+++ lib/apiScenario/postmanCollectionRunnerClient.ts
@@ -1,3 +1,4 @@
+import * as path from "path";
 import {
   Collection,
   Item,
@@ -11,8 +12,10 @@ import {
   Variable,
   VariableScope,
 } from "postman-collection";
+import { urlParse } from "@azure-tools/openapi-tools-common";
 import { xmsLongRunningOperation, xmsSkipUrlEncoding } from "../util/constants";
 import { JsonLoader } from "../swagger/jsonLoader";
+import { usePseudoRandom } from "../util/utils";
 import {
   ApiScenarioClientRequest,
   ApiScenarioRunnerClient,
@@ -38,9 +41,11 @@ import * as PostmanHelper from "./postmanHelper";
 import { VariableEnv } from "./variableEnv";
 
 export interface PostmanCollectionRunnerClientOption {
+  scenarioFile: string;
   collectionName?: string;
   runId: string;
   testProxy?: string;
+  testProxyAssets?: string;
   verbose?: boolean;
   skipAuth?: boolean;
   skipArmCall?: boolean;
@@ -245,6 +250,14 @@ export class PostmanCollectionRunnerClient implements ApiScenarioRunnerClient {
   }
 
   private startTestProxyRecording() {
+    const startRecordRequest: any = {
+      "x-recording-file": `${path.dirname(this.opts.scenarioFile)}/recordings/${
+        this.opts.collectionName
+      }.json`,
+    };
+    if (this.opts.testProxyAssets) {
+      startRecordRequest["x-recording-assets-file"] = this.opts.testProxyAssets;
+    }
     const { item } = this.addNewItem("Prepare", {
       name: "startTestProxyRecording",
       request: {
@@ -252,10 +265,13 @@ export class PostmanCollectionRunnerClient implements ApiScenarioRunnerClient {
         method: "POST",
         body: {
           mode: "raw",
-          raw: `{"x-recording-file": "./recordings/${this.opts.collectionName}_${this.opts.runId}.json"}`,
+          raw: `${JSON.stringify(startRecordRequest, null, 2)}`,
         },
       },
     });
+
+    item.request.addHeader({ key: "Content-Type", value: "application/json" });
+
     PostmanHelper.addEvent(
       item.events,
       "test",
@@ -274,8 +290,26 @@ pm.test("Started TestProxy recording", function() {
       request: {
         url: `${this.opts.testProxy}/record/stop`,
         method: "POST",
+        body: {
+          mode: "raw",
+          raw: `${JSON.stringify(
+            {
+              TIMESTAMP: "{{$isoTimestamp}}",
+              scenarioFile: this.opts.scenarioFile,
+              armEndpoint: this.runtimeEnv.get("armEndpoint"),
+              subscriptionId: this.runtimeEnv.get("subscriptionId"),
+              resourceGroup: this.runtimeEnv.get("resourceGroupName"),
+              location: this.runtimeEnv.get("location"),
+              runId: this.opts.runId,
+              randomSeed: usePseudoRandom.seed.toString(),
+            },
+            null,
+            2
+          )}`,
+        },
       },
     });
+    item.request.addHeader({ key: "Content-Type", value: "application/json" });
     item.request.addHeader({
       key: "x-recording-id",
       value: "{{x_recording_id}}",
@@ -452,6 +486,14 @@ pm.test("Stopped TestProxy recording", function() {
     }
   }
 
+  private resolveFilePath(filePath: string): string {
+    const url = urlParse(filePath);
+    if (url) {
+      throw new Error(`File path should be a local file path but got ${filePath}`);
+    }
+    return path.resolve(this.opts.scenarioFile, "..", filePath);
+  }
+
   public async sendRestCallRequest(
     clientRequest: ApiScenarioClientRequest,
     step: StepRestCall,
@@ -472,6 +514,30 @@ pm.test("Stopped TestProxy recording", function() {
             ? {
                 mode: "raw",
                 raw: JSON.stringify(convertPostmanFormat(clientRequest.body), null, 2),
+              }
+            : clientRequest.formData
+            ? {
+                mode: "formdata",
+                formdata: Object.entries(clientRequest.formData).map(([key, value]) => {
+                  if (value.type === "file") {
+                    return {
+                      key,
+                      type: "file",
+                      src: this.resolveFilePath(value.value),
+                    };
+                  } else {
+                    return {
+                      key,
+                      type: value.type,
+                      value: value.value,
+                    };
+                  }
+                }),
+              }
+            : clientRequest.file
+            ? {
+                mode: "file",
+                file: { src: this.resolveFilePath(clientRequest.file) },
               }
             : undefined,
         },