+++ lib/liveValidation/liveValidator.ts
@@ -254,7 +254,9 @@ export class LiveValidator {
     } catch (e) {
       // keeps building validator if it fails to tranform specs coz global transformers catches the exceptions and continue other schema transformings;
       // this error will be reported in validator building or validation runtime.
-      const errMsg = `Failed to transform loaded specs, detail error message:${e?.message}.ErrorStack:${e?.stack}`;
+      const errMsg = `Failed to transform loaded specs, detail error message:${
+        (e as any)?.message
+      }.\nError stack:${(e as any)?.stack}`;
       this.logging(
         errMsg,
         LiveValidatorLoggingLevels.error,
@@ -292,9 +294,10 @@ export class LiveValidator {
             }
           );
         } catch (e) {
-          const errMsg = `ErrorMessage:${e?.message}.ErrorStack:${e?.stack}`;
           this.logging(
-            `Failed to build validator for spec ${spec?._filePath}. ${errMsg}`,
+            `Failed to build validator for spec ${spec?._filePath}.\nErrorMessage:${
+              (e as any)?.message
+            }.\nErrorStack:${(e as any)?.stack}`,
             LiveValidatorLoggingLevels.error,
             LiveValidatorLoggingTypes.specTrace,
             "Oav.liveValidator.initialize.loader.buildAjvValidator",
@@ -380,9 +383,10 @@ export class LiveValidator {
           }
         );
       } catch (e) {
-        const errMsg = `ErrorMessage:${e?.message}.ErrorStack:${e?.stack}`;
         this.logging(
-          `Failed to build validator for spec ${spec?._filePath}. ${errMsg}`,
+          `Failed to build validator for spec ${spec?._filePath}.\nErrorMessage:${
+            (e as any)?.message
+          }.\nErrorStack:${(e as any)?.stack}`,
           LiveValidatorLoggingLevels.error,
           LiveValidatorLoggingTypes.specTrace,
           "Oav.liveValidator.loadAllSpecValidatorInBackground",
@@ -423,14 +427,7 @@ export class LiveValidator {
     operationInfo?: OperationContext
   ): Promise<LiveValidationResult> {
     const startTime = Date.now();
-    const correlationId = liveRequest.headers?.["x-ms-correlation-request-id"] || "";
-    const activityId = liveRequest.headers?.["x-ms-request-id"] || "";
-    const { info, error } = this.getOperationInfo(
-      liveRequest,
-      correlationId,
-      activityId,
-      operationInfo
-    );
+    const { info, error } = this.getOperationInfo(liveRequest, operationInfo);
     if (error !== undefined) {
       this.logging(
         `ErrorMessage:${error.message}.ErrorStack:${error.stack}`,
@@ -508,14 +505,7 @@ export class LiveValidator {
     operationInfo?: OperationContext
   ): Promise<LiveValidationResult> {
     const startTime = Date.now();
-    const correlationId = liveResponse.headers?.["x-ms-correlation-request-id"] || "";
-    const activityId = liveResponse.headers?.["x-ms-request-id"] || "";
-    const { info, error } = this.getOperationInfo(
-      specOperation,
-      correlationId,
-      activityId,
-      operationInfo
-    );
+    const { info, error } = this.getOperationInfo(specOperation, operationInfo);
     if (error !== undefined) {
       this.logging(
         `ErrorMessage:${error.message}.ErrorStack:${error.stack}`,
@@ -669,9 +659,7 @@ export class LiveValidator {
   }
 
   public getOperationInfo(
-    request: { url: string; method: string },
-    correlationId: string,
-    activityId: string,
+    request: { url: string; method: string; headers?: { [propertyName: string]: string } },
     operationInfo?: OperationContext
   ): {
     info: OperationContext;
@@ -683,11 +671,10 @@ export class LiveValidator {
     };
     try {
       if (info.validationRequest === undefined) {
-        info.validationRequest = this.parseValidationRequest(
+        info.validationRequest = parseValidationRequest(
           request.url,
           request.method,
-          correlationId,
-          activityId
+          request.headers
         );
       }
       if (info.operationMatch === undefined) {
@@ -700,15 +687,6 @@ export class LiveValidator {
     } catch (error) {
       return { info, error };
     }
-  }
-
-  public parseValidationRequest(
-    requestUrl: string,
-    requestMethod: string | undefined | null,
-    correlationId: string,
-    activityId: string
-  ): ValidationRequest {
-    return parseValidationRequest(requestUrl, requestMethod, correlationId, activityId);
   }
 
   private async getMatchedPaths(jsonsPattern: string | string[]): Promise<string[]> {
@@ -769,14 +747,7 @@ export class LiveValidator {
         runtimeException: runtimeException,
       };
     }
-    const correlationId =
-      requestResponseObj.liveRequest.headers?.["x-ms-correlation-request-id"] || "";
-    const activityId = requestResponseObj.liveRequest.headers?.["x-ms-request-id"] || "";
-    const { info, error } = this.getOperationInfo(
-      requestResponseObj.liveRequest,
-      correlationId,
-      activityId
-    );
+    const { info, error } = this.getOperationInfo(requestResponseObj.liveRequest);
     if (error !== undefined) {
       this.logging(
         `ErrorMessage:${error.message}.ErrorStack:${error.stack}`,
@@ -922,7 +893,9 @@ export class LiveValidator {
     } catch (err) {
       this.logging(
         `Unable to initialize "${swaggerPath}" file from SpecValidator. We are ` +
-          `ignoring this swagger file and continuing to build cache for other valid specs. ErrorMessage: ${err?.message};ErrorStack: ${err?.stack}`,
+          `ignoring this swagger file and continuing to build cache for other valid specs.\nErrorMessage: ${
+            (err as any)?.message
+          };\nErrorStack: ${(err as any)?.stack}`,
         LiveValidatorLoggingLevels.warn,
         LiveValidatorLoggingTypes.error
       );
@@ -1002,8 +975,10 @@ export function formatUrlToExpectedFormat(requestUrl: string): string {
  * @param activityId The id maps to request id, used by RPaaS.
  *
  * @returns parsed ValidationRequest info.
+ *
+ * @deprecated use parseValidationRequest instead.
  */
-export const parseValidationRequest = (
+export const legacyParseValidationRequest = (
   requestUrl: string,
   requestMethod: string | undefined | null,
   correlationId: string,
@@ -1034,6 +1009,28 @@ export const parseValidationRequest = (
     const e = new models.LiveValidationError(C.ErrorCodes.PotentialOperationSearchError.name, msg);
     throw e;
   }
+  return parseValidationRequest(requestUrl, requestMethod, {
+    "x-ms-correlation-request-id": correlationId,
+    "x-ms-request-id": activityId,
+  });
+};
+
+/**
+ * Parse the validation request information.
+ *
+ * @param requestUrl The url of service api call.
+ *
+ * @param requestMethod The http verb for the method to be used for lookup.
+ *
+ * @param headers Optional headers
+ *
+ * @returns parsed ValidationRequest info.
+ */
+export const parseValidationRequest = (
+  requestUrl: string,
+  requestMethod: string,
+  headers?: { [propertyName: string]: string }
+): ValidationRequest => {
   let queryStr;
   let apiVersion = "";
   let resourceType = "";
@@ -1070,8 +1067,8 @@ export const parseValidationRequest = (
     host: parsedUrl.host!,
     pathStr,
     query: queryStr,
-    correlationId,
-    activityId,
-    requestUrl,
+    requestUrl: requestUrl,
+    correlationId: headers?.["x-ms-correlation-request-id"] || "",
+    activityId: headers?.["x-ms-request-id"] || "",
   };
 };