+++ lib/util/utils.ts
@@ -140,6 +140,23 @@ export function checkAndResolveGithubUrl(inputPath: string): string {
   return inputPath;
 }
 
+/**
+ * Finds the git root directory for the given directory.
+ */
+export function findGitRootDirectory(dir: string): string | undefined {
+  while (true) {
+    const gitDir = path.join(dir, ".git");
+    if (fs.existsSync(gitDir)) {
+      return dir;
+    }
+    const newDIr = path.dirname(dir);
+    if (newDIr === dir) {
+      return undefined;
+    }
+    dir = newDIr;
+  }
+}
+
 /*
  * Merges source object into the target object
  * @param {object} source The object that needs to be merged
@@ -680,10 +697,8 @@ export const shuffleArray = (a: any[]) => {
   return a;
 };
 
-// eslint-disable-next-line prefer-const
-export let usePsudorandom = {
-  flag: false,
-  seed: 0,
+export const usePseudoRandom = {
+  seed: Math.floor(Math.random() * 10000000000),
 };
 
 /**
@@ -700,17 +715,14 @@ function* mulberry32(seed: number) {
 
 let generator: any = undefined;
 
-export const resetPsuedoRandomSeed = (seed?: number) => {
-  usePsudorandom.seed = seed ?? 0;
+export const resetPseudoRandomSeed = (seed?: number) => {
+  usePseudoRandom.seed = seed ?? Math.floor(Math.random() * 10000000000);
   generator = undefined;
 };
 
 export const getRandomString = (length?: number) => {
   if (generator === undefined) {
-    if (!usePsudorandom.flag) {
-      usePsudorandom.seed = Math.floor(Math.random() * 100000);
-    }
-    generator = mulberry32(usePsudorandom.seed);
+    generator = mulberry32(usePseudoRandom.seed);
   }
   return generator
     .next()